rename hello to diagnostics
