+++ ros2diagnostics_cli/ros2diagnostics_cli/api/__init__.py
@@ -1,0 +1,5 @@
+def get_hello_world():
+    return 'Hello, ROS 2 World!'
+
+def get_hello_world_leet():
+    return 'He110, R0S 2 W04ld!'
